+++ R/interesting-indices.r
@@ -128,6 +128,25 @@ cmass <- function() {
   }
 }
 
+#' Skewness index.
+#'
+#' Calculates the skewness index. See Cook, Buja and Cabrera (1993)
+#' Projection pursuit indexes based on orthonormal function expansions
+#'  for equations.
+#'
+#' @keywords hplot
+#' @export
+skewness <- function() {
+  function(mat) {
+    n <- nrow(mat)
+    d <- ncol(mat)
+
+    idx <- mean(rowSums(mat * exp(-0.5 * mat^2)))
+
+    idx
+  }
+}
+
 #' LDA projection pursuit index.
 #'
 #' Calculate the LDA projection pursuit index.  See Cook and Swayne (2007)